WIP Scaling optimisation
With a max of 28 web containers, and a max of 7 instances the service can be hit but 10 concurrent requests every 1-3 seconds and have enough resources to deploy at the same time.

20 web containers can create a max of 100 puma threads, 100 active record connections and 100 redis connections. This does not include the 6 worker containers that will run one per instance that add a max of 30 redis and 30 active record connections.

Requests at this load take 1-3 seconds for the home page (which does need caching too).

Desired usage should be 3 instances and 6 web containers which provides 30 puma threads, 30 database connections and 30 redis connections concurrently for usage.

- set logspout to daemon as it wasn’t always going onto autoscaled instances, sitting at 2/6 which won’t give us logs.
- deployments cannot introduce more than 20 containers into the pool, which can fail due to no more EC2 resources being available.