(feature) add view for awaiting feedback route
I'm not 100% happy with this as in order to get forms in table rows, I've has to use HTML5 form input attributes (https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#form). This is not supported in IE <= 11. We may be able to work around it with polyfills or JS